# telehealth-bangladesh/API's/telehealth_openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealNSight Telehealth Bangladesh API",
    "description": "Complete OpenAPI 3.0.3 specification covering all 58 endpoints across Django REST Framework, Node.js Auth Microservice, and FastAPI Worker.",
    "version": "1.0.0",
    "contact": {
      "name": "HealNSight Technical Team",
      "url": "https://github.com/infectedsourav/telehealth-bangladesh"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Django REST Backend (Core Clinical & Operations)"
    },
    {
      "url": "http://localhost:5000",
      "description": "Node.js Auth & OTP Microservice"
    },
    {
      "url": "http://localhost:6000",
      "description": "FastAPI AI & Vision Worker"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (obtainable from /api/login/ or /api/auth/login)."
      }
    }
  },
  "paths": {
...
    "/api/vision/analyze": {
      "post": {
        "tags": [
          "FastAPI Worker"
        ],
        "summary": "Multimodal ECG/CBC Chart Extraction",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "patient": 1,
                  "image_name": "ecg_scan.png",
                  "image_file": "BASE64"
                }
              }
            }
          }
        }
      }
    }
  }
}